envs/aisaac/wrapper.py
from __future__ import annotations
from typing import Any, Dict, List, Tuple

# Placeholder for a custom, procedurally generated multi-agent environment.
# Replace the below with actual imports once the AIsaac package is available.
try:
    import aisaac_env  # type: ignore
except Exception:  # pragma: no cover - optional dependency
    aisaac_env = None  # type: ignore

from envs import SocialEnvWrapper


class AIsaacWrapper(SocialEnvWrapper):
    """Stub wrapper for the custom AIsaac environment.

    Designed for true multi-agent from the start. The current implementation is
    a placeholder that documents the target API contract.
    """

    def __init__(self, num_agents: int = 2, seed: int | None = None) -> None:
        if aisaac_env is None:
            raise ImportError(
                "aisaac-env is not installed. Add it to requirements to use this wrapper."
            )
        self._agent_ids = [f"agent_{i}" for i in range(num_agents)]
        self._env = None  # real integration point
        self._seed = seed

    @property
    def agent_ids(self) -> List[str]:
        return self._agent_ids

    def reset(self, seed: int | None = None) -> Tuple[Dict[str, Any], Dict[str, Any]]:
        obs = {aid: None for aid in self._agent_ids}
        infos = {aid: {"note": "AIsaac environment not fully wired yet"} for aid in self._agent_ids}
        return obs, infos

    def step(
        self, actions: Dict[str, Any]
    ) -> Tuple[Dict[str, Any], Dict[str, float], Dict[str, bool], Dict[str, bool], Dict[str, Any]]:
        obs = {aid: None for aid in self._agent_ids}
        rewards = {aid: 0.0 for aid in self._agent_ids}
        terminated = {aid: False for aid in self._agent_ids}
        truncated = {aid: False for aid in self._agent_ids}
        infos = {aid: {"note": "AIsaac environment not fully wired yet"} for aid in self._agent_ids}
        return obs, rewards, terminated, truncated, infos

    def close(self) -> None:
        pass
