requirements.txt
# ==============================================================================
# CoffeeShop - Core Dependencies
# ==============================================================================

# -- Core Utilities --
numpy>=1.24.0
pyyaml>=6.0
rich>=13.7.0
loguru>=0.7.0
packaging>=24.0
tqdm>=4.66.0

# -- Config Management --
omegaconf>=2.3.0
# hydra-core>=1.3.0       # Uncomment if you adopt Hydra for YAML config routing

# -- PyTorch Ecosystem --
# Defaulting to standard PyPI (which defaults to CUDA on Linux/Windows).
# If you explicitly need CPU-only, uncomment the extra-index line below:
# --extra-index-url https://download.pytorch.org/whl/cpu

torch>=2.2.0
# torchvision>=0.17.0     # Optional: Uncomment if visual observations are used
# torchaudio>=2.2.0       # Optional

# -- Reinforcement Learning Ecosystem (Env-Agnostic Glue) --
gymnasium>=0.29.0
shimmy>=1.3.0             # Gym API adapters
# pettingzoo>=1.24.0      # Optional: Uncomment if using standard MARL baseline APIs

# -- Experiment Tracking --
wandb>=0.16.0             # Optional: Default weights and biases tracker
# tensorboard>=2.15.0     # Optional: Alternative tracker

# ==============================================================================
# Environment-Specific (Enable as needed based on benchmark)
# ==============================================================================


# -- Overcooked-AI --
overcooked-ai>=1.1.0
pygame>=2.5.0

# -- Crafter Benchmark --
crafter>=1.8.0

# -- NetHack Learning Environment (NLE) --
nle==0.9.1              # Note: May require cmake and gcc/build-essential installed on OS

# -- Google Research Football (Phase 2 Benchmark) --
# gfootball @ git+https://github.com/google-research/football.git